+++ Sources/StreamChatUI/ChatChannel/ChatChannelVC.swift
@@ -233,7 +233,7 @@ open class ChatChannelVC: _ViewController,
     /// Marks the channel read and updates the UI optimistically.
     public func markRead() {
         channelController.markRead()
-        messageListVC.scrollToLatestMessageButton.content = .noUnread
+        messageListVC.scrollToBottomButton.content = .noUnread
     }
 
     /// Jump to a given message.
@@ -415,7 +415,7 @@ open class ChatChannelVC: _ViewController,
         didUpdateChannel channel: EntityChange<ChatChannel>
     ) {
         let channelUnreadCount = channelController.channel?.unreadCount ?? .noUnread
-        messageListVC.scrollToLatestMessageButton.content = channelUnreadCount
+        messageListVC.scrollToBottomButton.content = channelUnreadCount
 
         guard channelController.firstUnreadMessageId != firstUnreadMessageId else { return }
         let previousUnreadMessageId = firstUnreadMessageId