+++ Sources/StreamChatUI/ChatMessageList/ChatMessageListVC.swift
@@ -74,8 +74,8 @@ open class ChatMessageListVC: _ViewController,
 
     /// A button to scroll the collection view to the bottom.
     /// Visible when there is unread message and the collection view is not at the bottom already.
-    open private(set) lazy var scrollToLatestMessageButton: ScrollToLatestMessageButton = components
-        .scrollToLatestMessageButton
+    open private(set) lazy var scrollToBottomButton: ScrollToBottomButton = components
+        .scrollToBottomButton
         .init()
         .withoutAutoresizingMaskConstraints
 
@@ -153,7 +153,7 @@ open class ChatMessageListVC: _ViewController,
         panGestureRecognizer.delegate = self
         listView.addGestureRecognizer(panGestureRecognizer)
 
-        scrollToLatestMessageButton.addTarget(self, action: #selector(scrollToLatestMessage), for: .touchUpInside)
+        scrollToBottomButton.addTarget(self, action: #selector(didTapScrollToBottomButton), for: .touchUpInside)
     }
 
     override open func setUpLayout() {
@@ -171,12 +171,12 @@ open class ChatMessageListVC: _ViewController,
         typingIndicatorView.pin(anchors: [.leading, .trailing], to: view)
         typingIndicatorView.bottomAnchor.pin(equalTo: listView.bottomAnchor).isActive = true
 
-        view.addSubview(scrollToLatestMessageButton)
-        listView.bottomAnchor.pin(equalToSystemSpacingBelow: scrollToLatestMessageButton.bottomAnchor).isActive = true
-        scrollToLatestMessageButton.trailingAnchor.pin(equalTo: view.layoutMarginsGuide.trailingAnchor).isActive = true
-        scrollToLatestMessageButton.widthAnchor.pin(equalTo: scrollToLatestMessageButton.heightAnchor).isActive = true
-        scrollToLatestMessageButton.heightAnchor.pin(equalToConstant: 40).isActive = true
-        scrollToLatestMessageButton.isHidden = true
+        view.addSubview(scrollToBottomButton)
+        listView.bottomAnchor.pin(equalToSystemSpacingBelow: scrollToBottomButton.bottomAnchor).isActive = true
+        scrollToBottomButton.trailingAnchor.pin(equalTo: view.layoutMarginsGuide.trailingAnchor).isActive = true
+        scrollToBottomButton.widthAnchor.pin(equalTo: scrollToBottomButton.heightAnchor).isActive = true
+        scrollToBottomButton.heightAnchor.pin(equalToConstant: 40).isActive = true
+        scrollToBottomButton.isHidden = true
 
         if isDateOverlayEnabled {
             view.addSubview(dateOverlayView)
@@ -240,36 +240,35 @@ open class ChatMessageListVC: _ViewController,
         )
     }
 
-    /// Set the visibility of `scrollToLatestMessageButton`.
-    @available(*, deprecated, message: "use updateScrollToBottomButtonVisibility(animated:) instead.")
-    open func setScrollToLatestMessageButton(visible: Bool, animated: Bool = true) {
-        updateScrollToBottomButtonVisibility()
-    }
-
-    /// Set the visibility of `scrollToLatestMessageButton`.
+    /// Set the visibility of `scrollToBottomButton`.
     open func updateScrollToBottomButtonVisibility(animated: Bool = true) {
         let isVisible = isScrollToBottomButtonVisible
-        if isVisible { scrollToLatestMessageButton.isVisible = true }
+        if isVisible { scrollToBottomButton.isVisible = true }
         Animate(isAnimated: animated, {
-            self.scrollToLatestMessageButton.alpha = isVisible ? 1 : 0
+            self.scrollToBottomButton.alpha = isVisible ? 1 : 0
         }, completion: { _ in
-            if !isVisible { self.scrollToLatestMessageButton.isVisible = false }
+            if !isVisible { self.scrollToBottomButton.isVisible = false }
         })
     }
 
-    /// Action for `scrollToLatestMessageButton` that scroll to most recent message.
-    @objc open func scrollToLatestMessage() {
+    /// Action for `scrollToBottomButton` that scroll to most recent message.
+    @objc open func didTapScrollToBottomButton() {
         guard isFirstPageLoaded else {
             jumpToFirstPage()
             return
         }
 
-        scrollToMostRecentMessage()
+        scrollToBottom()
     }
 
-    /// Scrolls to most recent message
-    open func scrollToMostRecentMessage(animated: Bool = true) {
-        listView.scrollToMostRecentMessage(animated: animated)
+    /// Scroll to the bottom of the message list.
+    open func scrollToBottom(animated: Bool = true) {
+        listView.scrollToBottom(animated: animated)
+    }
+
+    /// Scroll to the top of the message list.
+    open func scrollToTop(animated: Bool = true) {
+        listView.scrollToTop(animated: animated)
     }
 
     func updateUnreadMessagesSeparator(at id: MessageId?, previousId: MessageId?) {
@@ -564,7 +563,7 @@ open class ChatMessageListVC: _ViewController,
     /// Jump to the first page of the message list.
     internal func jumpToFirstPage() {
         delegate?.chatMessageListVCShouldLoadFirstPage(self)
-        scrollToLatestMessageButton.isHidden = true
+        scrollToBottomButton.isHidden = true
         listView.reloadSkippedMessages()
     }
 
@@ -879,6 +878,40 @@ open class ChatMessageListVC: _ViewController,
         audioSessionFeedbackGenerator.feedbackForSeeking()
         audioPlayer?.seek(to: timeInterval)
     }
+
+    // MARK: - Deprecations
+
+    /// Jump to a given message.
+    /// In case the message is already loaded, it directly goes to it.
+    /// If not, it will load the messages around it and go to that page.
+    ///
+    /// - Parameter message: The message which the message list should go to.
+    /// - Parameter onHighlight: An optional closure to provide highlighting style when the message appears on screen.
+    @available(*, deprecated, renamed: "jumpToMessage(id:onHighlight:)")
+    public func jumpToMessage(_ message: ChatMessage, onHighlight: ((IndexPath) -> Void)? = nil) {
+        jumpToMessage(id: message.id, onHighlight: onHighlight)
+    }
+
+    /// Set the visibility of `scrollToBottomButton`.
+    @available(*, deprecated, message: "use updateScrollToBottomButtonVisibility(animated:) instead.")
+    open func setScrollToLatestMessageButton(visible: Bool, animated: Bool = true) {
+        updateScrollToBottomButtonVisibility()
+    }
+
+    @available(*, deprecated, renamed: "scrollToBottom(animated:)")
+    open func scrollToMostRecentMessage(animated: Bool = true) {
+        listView.scrollToBottom(animated: animated)
+    }
+
+    @available(*, deprecated, renamed: "scrollToBottomButton")
+    open var scrollToLatestMessageButton: ScrollToBottomButton {
+        scrollToBottomButton
+    }
+
+    @available(*, deprecated, renamed: "didTapScrollToBottomButton")
+    @objc open func scrollToLatestMessage() {
+        didTapScrollToBottomButton()
+    }
 }
 
 // MARK: - Handle Message Updates
@@ -903,7 +936,7 @@ private extension ChatMessageListVC {
             self?.updateScrollToBottomButtonVisibility()
 
             UIView.performWithoutAnimation {
-                self?.scrollToMostRecentMessageIfNeeded(with: changes, newestChange: newestChange)
+                self?.scrollToBottomIfNeeded(with: changes, newestChange: newestChange)
                 self?.reloadMovedMessage(newestChange: newestChange)
                 self?.reloadPreviousMessagesForVisibleRemoves(with: changes)
                 self?.reloadPreviousMessageWhenInsertingNewMessage()
@@ -980,12 +1013,12 @@ private extension ChatMessageListVC {
 
     // Scroll to the bottom if the new message was sent by
     // the current user, or moved by the current user, and the first page is loaded.
-    func scrollToMostRecentMessageIfNeeded(with changes: [ListChange<ChatMessage>], newestChange: ListChange<ChatMessage>?) {
+    func scrollToBottomIfNeeded(with changes: [ListChange<ChatMessage>], newestChange: ListChange<ChatMessage>?) {
         guard isFirstPageLoaded else { return }
         guard let newMessage = newestChange?.item else { return }
         let newestChangeIsInsertionOrMove = newestChange?.isInsertion == true || newestChange?.isMove == true
         if newestChangeIsInsertionOrMove && newMessage.isSentByCurrentUser {
-            scrollToMostRecentMessage()
+            scrollToBottom()
         }
     }
 
@@ -996,20 +1029,5 @@ private extension ChatMessageListVC {
             let movedIndexPath = IndexPath(item: 0, section: 0)
             listView.reloadRows(at: [movedIndexPath], with: .none)
         }
-    }
-}
-
-// MARK: - Deprecations
-
-extension ChatMessageListVC {
-    /// Jump to a given message.
-    /// In case the message is already loaded, it directly goes to it.
-    /// If not, it will load the messages around it and go to that page.
-    ///
-    /// - Parameter message: The message which the message list should go to.
-    /// - Parameter onHighlight: An optional closure to provide highlighting style when the message appears on screen.
-    @available(*, deprecated, renamed: "jumpToMessage(id:onHighlight:)")
-    public func jumpToMessage(_ message: ChatMessage, onHighlight: ((IndexPath) -> Void)? = nil) {
-        jumpToMessage(id: message.id, onHighlight: onHighlight)
     }
 }